[FIX] (#28) - change post controller's annotation and refectory RequestWrapper
