+++ server/src/main/java/com/iieee/server/filter/wrapper/CachedBodyServletInputStream.java
@@ -1,0 +1,40 @@
+package com.iieee.server.filter.wrapper;
+
+import javax.servlet.ReadListener;
+import javax.servlet.ServletInputStream;
+import java.io.ByteArrayInputStream;
+import java.io.IOException;
+import java.io.InputStream;
+
+public class CachedBodyServletInputStream extends ServletInputStream {
+    private final InputStream cachedBodyInputStream;
+
+    public CachedBodyServletInputStream(byte[] cachedBody) {
+        this.cachedBodyInputStream = new ByteArrayInputStream(cachedBody);
+    }
+
+    @Override
+    public boolean isFinished() {
+        try {
+            return cachedBodyInputStream.available() == 0;
+        } catch (IOException e) {
+            e.printStackTrace();
+        }
+        return false;
+    }
+
+    @Override
+    public boolean isReady() {
+        return true;
+    }
+
+    @Override
+    public void setReadListener(ReadListener readListener) {
+        throw new UnsupportedOperationException();
+    }
+
+    @Override
+    public int read() throws IOException {
+        return cachedBodyInputStream.read();
+    }
+}
