+++ server/src/main/java/com/iieee/server/filter/wrapper/RequestWrapper.java
@@ -2,52 +2,34 @@ package com.iieee.server.filter.wrapper;
 
 import org.springframework.util.StreamUtils;
 
-import javax.servlet.ReadListener;
 import javax.servlet.ServletInputStream;
+import javax.servlet.ServletRequest;
 import javax.servlet.http.HttpServletRequest;
 import javax.servlet.http.HttpServletRequestWrapper;
-import java.io.ByteArrayInputStream;
-import java.io.IOException;
-import java.io.InputStream;
+import java.io.*;
 
 public class RequestWrapper extends HttpServletRequestWrapper {
-    private final byte[] cachedInputStream;
+    private final byte[] cachedBody;
 
     public RequestWrapper(HttpServletRequest request) throws IOException {
         super(request);
         InputStream requestInputStream = request.getInputStream();
-        this.cachedInputStream = StreamUtils.copyToByteArray(requestInputStream);
+        this.cachedBody = StreamUtils.copyToByteArray(requestInputStream);
     }
 
     @Override
-    public ServletInputStream getInputStream() {
-        return new ServletInputStream() {
-            private final InputStream cachedBodyInputStream = new ByteArrayInputStream(cachedInputStream);
+    public ServletInputStream getInputStream() throws IOException {
+        return new CachedBodyServletInputStream(this.cachedBody);
+    }
 
-            @Override
-            public boolean isFinished() {
-                try {
-                    return cachedBodyInputStream.available() == 0;
-                } catch (IOException e) {
-                    e.printStackTrace();
-                }
-                return false;
-            }
+    @Override
+    public BufferedReader getReader() throws IOException {
+        ByteArrayInputStream byteArrayInputStream = new ByteArrayInputStream(this.cachedBody);
+        return new BufferedReader(new InputStreamReader(byteArrayInputStream));
+    }
 
-            @Override
-            public boolean isReady() {
-                return true;
-            }
-
-            @Override
-            public void setReadListener(ReadListener listener) {
-                throw new UnsupportedOperationException();
-            }
-
-            @Override
-            public int read() throws IOException {
-                return cachedBodyInputStream.read();
-            }
-        };
+    @Override
+    public ServletRequest getRequest() {
+        return super.getRequest();
     }
 }
